Bonus code snippet available !
This script connects to the NordVPN API, asks for your username and password, and gets a list of proxy servers. It saves this list in two ways: one as a simple list in `socks_ips.txt`, and another as full connection URLs in `socks_urls.txt` (ready for FoxyProxy). Enjoy !